requirements/production.txt
Django==1.8.19
psycopg2-binary==2.7.5
pycrypto==2.6.1
whitenoise==3.2.2
django-cors-headers==1.2.2
redis==2.10.6
django-redis==4.5.0
pytz==2016.7
django-rq==0.9.4

dj-database-url==0.4.1
django-recaptcha==1.2.0
django-jsonfield==1.0.1
python-dotenv==0.10.2
rq==0.12.0
mock==2.0.0
freezegun==0.3.12
supervisor==4.2.5
Twisted==16.6.0
channels==1.1.8
asgi_redis==1.4.3
beatserver==0.0.4
